+++ src/api/index.ts
@@ -5,6 +5,6 @@ export const kyInstance = ky.create({
   headers: {
     "Content-Type": "application/json",
     Authorization:
-      "Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ0ZXN0QGVtYWlsLmNvbSIsImF1dGgiOiJST0xFX1VTRVIiLCJleHAiOjE3MzIxMTM5MDN9.jwoW5QGBUlE1bvmPUVBTZxj-b_M6k1tNPt3xGqTYhEA",
+      "Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ0ZXN0QGVtYWlsLmNvbSIsImF1dGgiOiJST0xFX1VTRVIiLCJleHAiOjE3MzIyMDA4NjN9.oDMOpF9JtmBO01UHRzSu58s-bveE5Ou_jisDq9Eah5Q",
   },
 });
