+++ src/view/group/dashboard/NoticeModal/NoticeList/index.css.ts
@@ -61,16 +61,6 @@ export const avatarStyle = style({
   height: "4rem",
 });
 
-export const buttonStyle = style({
-  float: "right",
-  transform: "translate(-2rem, 0)",
-  width: "8.4rem",
-  height: "4rem",
-
-  ...theme.font.Body3_SB_14,
-  color: theme.color.lg2,
-});
-
 export const paginationStyle = style({
   paddingTop: "1rem",
 });
